requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0
tqdm>=4.65.0
requests>=2.31.0
matplotlib>=3.8.0
scipy>=1.11.0
